no automatic add, major refactors
