+++ src/pages/PoolFinder/index.tsx
@@ -1,5 +1,5 @@
 import { JSBI, Pair, Token, TokenAmount } from '@uniswap/sdk'
-import React, { useEffect, useState } from 'react'
+import React, { useCallback, useEffect, useState } from 'react'
 import { Plus } from 'react-feather'
 import { RouteComponentProps } from 'react-router-dom'
 import { Text } from 'rebass'
@@ -8,7 +8,7 @@ import { LightCard } from '../../components/Card'
 import { AutoColumn, ColumnCenter } from '../../components/Column'
 import PositionCard from '../../components/PositionCard'
 import Row from '../../components/Row'
-import SearchModal from '../../components/SearchModal'
+import TokenSearchModal from '../../components/SearchModal/TokenSearchModal'
 import TokenLogo from '../../components/TokenLogo'
 import { usePair } from '../../data/Reserves'
 import { useActiveWeb3React } from '../../hooks'
@@ -48,6 +48,17 @@ export default function PoolFinder({ history }: RouteComponentProps) {
     pair === null ||
     (!!pair && JSBI.equal(pair.reserve0.raw, JSBI.BigInt(0)) && JSBI.equal(pair.reserve1.raw, JSBI.BigInt(0)))
   const allowImport: boolean = position && JSBI.greaterThan(position.raw, JSBI.BigInt(0))
+
+  const handleTokenSelect = useCallback(
+    (address: string) => {
+      activeField === Fields.TOKEN0 ? setToken0Address(address) : setToken1Address(address)
+    },
+    [activeField]
+  )
+
+  const handleSearchDismiss = useCallback(() => {
+    setShowSearch(false)
+  }, [setShowSearch])
 
   return (
     <AppBody>
@@ -146,15 +157,10 @@ export default function PoolFinder({ history }: RouteComponentProps) {
           </Text>
         </ButtonPrimary>
       </AutoColumn>
-      <SearchModal
+      <TokenSearchModal
         isOpen={showSearch}
-        filterType="tokens"
-        onTokenSelect={address => {
-          activeField === Fields.TOKEN0 ? setToken0Address(address) : setToken1Address(address)
-        }}
-        onDismiss={() => {
-          setShowSearch(false)
-        }}
+        onTokenSelect={handleTokenSelect}
+        onDismiss={handleSearchDismiss}
         hiddenToken={activeField === Fields.TOKEN0 ? token1Address : token0Address}
       />
     </AppBody>
